requirements.txt
pydantic
google-generativeai
python-dotenv
fastapi[standard]
loguru
uvicorn
